Deploy rbd-mirror using cephadm
This patch enables the support of the rbd-mirror deployment via
the spec. The approach is the same used for the other daemons and
when the OS::TripleO::Services::CephRbdMirror is included in a
TripleO role, the rbd-mirror tasks are executed (the spec is built
and applied against the Ceph cluster using the ceph_mkspec module).

Change-Id: I0d1ce33270ccf8636925c3534812c3b1cf650103
(cherry picked from commit 9b6e8a67b5994d12b931fb34e221b4223584b038)