More descriptive task names
As highlighted by marios on [1] and [2] the task names were
not descriptive of what it was being done, as it implied that some kind
of verification was being done.

This patch does not change any functionality, but it changes the names
of the tasks to something more according to what it is being done.

[1] https://review.opendev.org/c/openstack/tripleo-ansible/+/791450/12/tripleo_ansible/roles/backup_and_restore/tasks/ceph_authentication.yml
[2] https://review.opendev.org/c/openstack/tripleo-ansible/+/791450/12/tripleo_ansible/roles/backup_and_restore/tasks/db_backup.yml

Change-Id: I19445ffd1c8d35972fbf8ea3da4a9fbe599f6127
(cherry picked from commit 138f9f6708546cf17f28334be93ee2acc3fbf9a8)