Only write inventory network-config if "configure_networking"
When overcloud node provision command is ran with '--network-ports'
and not '--network-config', don't write the file with extra ansible
inventory network configuration.

When '--network-ports' is used the network configuration is done as
part of the deploy steps playbook, and the THT parameters should be
the source of truth.

Depends-On: Ief11590f3421ffd6fb1b4312c07ce9de48f17c3d
Change-Id: I0b8d96e40b0427458e977094f641f2f42ceaa199
(cherry picked from commit c6645a155d004cb1c7e80f4e5f19836550af6bce)