Make sure the _admin label is present on mons
If a ceph mon does not have the _admin label, then the
hosts ceph.conf may be deleted. All mons are considered
admin hosts, so apply this label to any host with the
mon service.
In addition, this change fixes the default alertmanager
port (it should be 9093 as 9094 is reserved for clustering
purposes) and the cluster_name is explicitly passed to
the mkspec tasks.

Change-Id: Ifaa37eebf0759ec8861de86266c0a0dd686cc6b3