Add playbook cli-overcloud-node-growvols.yaml
Just like cli-overcloud-node-network-config.yaml, this playbook will
be run by the "openstack overcloud node provision" directly after
provisioning so that whole disk images will have their /var volume
grown to 100%.

If the growvols utility is not found the playbook will halt without
error, because this will be an overcloud-full partition image which
grows the root partition with the growpart utility.

There will be a mechanism in the "openstack overcloud node provision"
command to specify different growvol arguments. This will not be
required for most nodes, but swift storage nodes will have to grow
both /srv and /var.

Change-Id: I0a847ad4077a02a02ad817dd189a6a31c3637a93
Blueprint: whole-disk-default