Container manage module, attempt 2
Use a module to manage podman containers rather than the async task
loops.  This change also includes fixes for container-puppet json data
to match the expectations of the module.

This changed the default state from created to started which matches
what used to occur in the tripleo_container_manage role. Additionally
an additional check was added to an exec to ensure the target container
is up before running the exec. We will sleep for 5 seconds and check
again if the container is not running.

Change-Id: I371fbb5d3489c70e2bd4e3473faaa3e65f591d63
(cherry picked from commit 72f42e0f3b870d435ce9083480dd189e163ef1c9)