Add tripleo_ceph_distribute_keys role
The nodes having an _admin label should be able to interact with
the ceph cluster using the ceph.client.admin keyring. This label
is typically added to the mon/mgr group, but they need the admin
keyring to run any ceph command (via cephadm shell or any other
client tool).
This change just adds a new role to properly distribute the admin
keyring right after the ceph cluster is deployed, and improves the
cephadm.yaml playbook to include this action.

Change-Id: Id17f69c714dec75ec48862f6f0cf900fae5e7c30