Restart octavia rsyslog after updating the config
After creating the octavia rsyslog configuration file, the
octavia_rsyslog container must be restarted if it's already running,
otherwise the new configuration is not applied in the container.

This issue is new, the octavia-ansible playbook was moved from step 5 to
post_deploy, which means that the file is now created after the first start
of the container.

Change-Id: Ibb439d47d551adc5af2d87234b732565af1cc90f