Remove hardcoded cloud_name and auth_type used
We use stackrc which sets OS_CLOUD env variable, when running
the cli playbooks. We've been removing hardcoded cloud_names
specified in playbooks to not conflict with it.

Also, 'auth_type' would be picked from clouds.yaml and not
required to be passed an argument.

For standalone ironic using noauth, clouds.yaml would look
like below and there is no need to have any logic in the
modules.

clouds:
  undercloud:
    auth: {}
    auth_type: none
    baremetal_api_version: '1'
    baremetal_endpoint_override: https://192.168.24.2:13385

Change-Id: Iacf16819ee98ed9d41abbe0e7f6a77f36e803de2