Copy generated /boot/grub2/grubenv to EFI dirs
There is an issue with grub2 in rhel-8.2 where generating grub.cfg in
an EFI directory will regenerate grubenv in /boot/grub2. Then when the
host is rebooted, the stale grubenv file in the EFI directory is used,
and any new kernal args are not used.

This change works around this by copying /boot/grub2/grubenv to the
EFI directories. Newer versions of grub2 many not show this behaviour
but the workaround can remain in place.

This change also adds /boot/efi/EFI/BOOT to the list of EFI
directories checked for, as it should have been in the original list.

See https://bugzilla.redhat.com/show_bug.cgi?id=1974507

Change-Id: I4f8664c9033502b25b40646cfd8976cc5638e643