Backup and Restore playbook is not 100% idempotent when running in snapshot mode
This fix a small issue that can happen when the user directly cancels
the execution of the playbook when the rear backup is happening. After
that, the node remains disconnected from pacemaker and it is necessary
to add it back to the cluster so the mysql-galera container is running
when we do the mysql dump.

As the pcs cluster start is an idempotent command in itself, we do not
need to perform a verification check before running it nor it will
generate errors (return 1) when ran if it was already running.

Closes-Bug: #1934897

Change-Id: I9a425665f4714751d8b1a0bd381533e4c383df9d
(cherry picked from commit aa93f9d7a05d33bdf0aa6177da10f51a1e784bc2)
(cherry picked from commit a22d8d9817e07b76ccd83211659fea55796b735b)