Explicitly set ceph_dashboard_vip when Ceph Dashboard is enabled
When the Ceph Dashboard environment file is added, the haproxy config
is created for both the ceph dashboard and the monitoring stack (made
by grafana, prometheus, alertmanager, node-exporter).
We used to have the 'ceph_dashboard_vip' rely to the ctlplane network
unless the StorageDashboard composable network is explicitly added to
the overcloud deployment.
This patch, instead, explicitly sets the ceph_dashboard_vip using the
'CephDashboardNetwork' parameter defined in tripleo-heat-templates.
By doing this the operator has the flexibility to use the network
defined via tht to reach the Ceph Dashboard, resulting in a more
consistent configuration according to the user expectations.

Change-Id: Icc68dad8074dd02d84f76efac45ed08b42a2bd2f