Allow for cron programming of backups
Normally, backups are not done interactively, but during nights or
weekends. This patch introduces the capacity of introducing in an easy
way cron programming that will execute backups in predictable intervals.

Not a clean backup due to a small conflict with the variables file.

Closes-Bug: #1935675

Change-Id: I86b3fb2c3d846bcd56859e0d572f1fe4d5e148fa
(cherry picked from commit 482545b07c8a92767b39a94320d445927626be56)