tripleo-podman: Enable and start podman.socket
It might be a good idea to control wether or not the podman.socket
service is enabled in case we want to podman --remote.

We plan on using this in the wrapper services instead of relying on
nsenter.

Cherry-pick included in this patch.
* https://review.opendev.org/c/openstack/tripleo-ansible/+/817202

Changes which are already cherry-picked:
* https://review.opendev.org/q/Ifca971e14abb184c4b7b8a0d004e8c2fbd0ecb68
* https://review.opendev.org/q/Iefe861f91cefe2a9cf773cae98b2440566ae8b5e
* https://review.opendev.org/q/Ibdf87463d4a86dd225a1f1b89326407628facb65

Change-Id: I52d63d34a9c01bac7e77f2ce9154aa719b4f82b5
(cherry picked from commit 7f5ecb62b5a353c657c4a0dc056caa35fbb583f7)