Improve the config-key settings when Ceph is deployed
The purpose of this patch is having the ability to have a general
set of tasks where the mgr (and global) config key are set.
For development and debug purposes the tripleo_cephadm_debug
parameter is added, and can be used to have a more verbose output
from the orchestrator.

Change-Id: Ic2e54e537b50cb92b1650e78f1af61703e877763