tripleo-kernel: prevent reboot when TSX flag was added
During the 16.2 upgrade path, we enforce the inclusion of tsx flag by
operators. If there was already kernelargs present, the node won't get
rebooted, but if it's a new KernelArgs, the node will get rebooted as
per the related bug.

We want to prevent the unexpected reboot of nodes with workload when
the TSX flag was added.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1975240
Change-Id: Ib6140c8991469c950fd7e1b323c614103394e3f1