Add support for v2 registries.conf
https://github.com/containers/image/blob/master/docs/containers-registries.conf.5.md#version-2

This allows for mirrors to be configured for specific hosts however it
is incompatibilty with the default v1 configuration so we have to nuke
the existing configuration.  Additionally it uses TOML which there is
currently no ansible module to manage.

Adds to_json to address issue seen with py2 jobs at [1]. The depends-on
is unrelated but needed to pass centos7 jobs in ci.

[1] https://review.opendev.org/c/openstack/tripleo-ansible/+/801615/2#message-821960bb5908358d27225a12fdc5c84ab7bfe0d0
Depends-On: https://review.opendev.org/c/openstack/tripleo-quickstart/+/802431
Related-Bug: 1936825
Change-Id: Ic35155f04bf05913b9e9b8eaa22fe6c02515396c
(cherry picked from commit bf80fe922b2d915675f01794e73893c018688916)