Set ganesha_vip when no StoragetNFS network
is deployed.

Normally when Ganesha is deployed for CephFS through NFS,
the StorageNFS isolated network is deployed and we set
the address for ganesha_vip on that network.  But with
Standalone role the StorageNFS network is not deployed
so ganesha_vip is not set in the tripleo hieradata unless
it is provided by extra config.

Set it to net_vip_map.external in the event that it has
not been set on the StorageNFS network.  Note that external
is chosen rather than ctlplane since the VIP must be reachable
from compute instances, guest VMs, that should have no access
to the ctlplane network.

Closes-Bug: #1911022
Change-Id: Ia97533c0b8b3b8498a173e1d363323dcdc77cce4
(cherry picked from commit 0aceca00e76e967ee9b5efc30da0bc3ee23acb2e)
(cherry picked from commit 520b1e1edb83d9f9e101695900667c80a4cdbc8f)