Enforce minimum Ceph client version to Mimic
This is so that we can take advantage of the deferred deletion
feature in Cinder. This change is done in Ansible instead of
TripleO Heat Templates so that it is deployed_ceph compatible
while still having an effect when deployed_ceph is not used.

Closes-Bug: #1938833
Change-Id: Ic76953f1a671ea0416da829acb37bd1d787a55dc
(cherry picked from commit bb5e1cfaab7f579037af9517d4be37e1beea5094)