Align tasks between config-download playbooks
cli-config-download.yaml and cli-config-download-export.yaml both
download the config from the stack. This commit aligns the tasks between
the 2 so that they both setup and download the config in the same way.

This allows a single playbook to be used by tripleoclient
(cli-config-download.yaml) as done in
Ia765c2c2ea548a1dcd20f592eda8741f40a1cf16

cli-config-download-export.yaml will be removed in a subsequent patch as
it will no longer be used by tripleoclient.

Change-Id: Ie47abe2d7c5d7891aceb88d9008d45d7922dea15
Signed-off-by: James Slagle <jslagle@redhat.com>