tripleo_frr: Allow a password to be set in FRR's neighbor section
Set a MD5 password to be used with the tcp socket that is being used to
connect to the remote peer. For more information, please refer to
https://docs.frrouting.org/en/latest/bgp.html#clicmd-neighbor-PEER-password-PASSWORD

Note that this is disabled by default. When enabled it won't work on a FIPS-enabled
system. FRR is already patched so that it will fail to start in such a
situation,

[1] https://git.centos.org/rpms/frr/blob/c8s/f/SOURCES/0004-fips-mode.patch

Change-Id: I0a13ebb127bb484804169d7e4930138bcb01d5a9
(cherry picked from commit 246bae8a227a21aa051e278658439806f01434d8)