Allow BFD timeouts to be customized
Especially in virtual environments the default rfc-specified BFD timeouts are quite
aggressive. A small overload on the control plane has the chances of
making us miss heartbeat and declaring a link down.

Let's add some parameters (set to the RFC-defined defaults) in order to
be able to customize them.

Tested with:
  ControllerRack1ExtraGroupVars:
    tripleo_frr_bfd_detect_multiplier: 10
    tripleo_frr_bfd_transmit_interval: 500
    tripleo_frr_bfd_receive_interval: 500

Obtained the following in the config:
router bgp 64999
  bgp router-id 172.30.1.1
  bgp log-neighbor-changes
  bgp graceful-shutdown
...
  neighbor uplink peer-group
  neighbor uplink bfd
  neighbor uplink bfd profile tripleo
...
bfd
  profile tripleo
    detect-multiplier 10
    transmit-interval 500
    receive-interval 500

vtysh confirms that the config is applied correctly:
[root@ctrl-1-0 ~]# podman exec -it frr vtysh -c 'show running-config' |grep -A4 bfd
router bgp 64999
 bgp router-id 172.30.1.1
 bgp log-neighbor-changes
--
   neighbor uplink bfd
 neighbor uplink bfd profile tripleo
 neighbor uplink ttl-security hops 1
 neighbor enp2s0 interface peer-group uplink
 neighbor enp3s0 interface peer-group uplink
 !
--
  bfd
 profile tripleo
  detect-multiplier 10
  transmit-interval 500
  receive-interval 500

Change-Id: I8ff745be02b476bcfeac1bfd162902d6fffdd6cb