Remove stack_action for os-net-config
Since we're using ephemeral Heat, there is no UPDATE
stack action. All actions are now CREATE. Therefore,
we can no longer rely on the stack action to update
the network configuration and NetworkDeploymentActions
is no longer an option. This change adds a new parameter
tripleo_network_config_update. This will be a bool.

Change-Id: If53577a26edcb2dc67458603286471dfb87f0bda