Check whether an HA resource already exists explicitly
With ephemeral heat we lost the meaning of the 'stack_action' hiera key
which we previously used to distinguish between fresh deployment and
pre-existing deployment (aka redeploy).
Since this hiera key is not available anymore, we simply check if the
resource existed before running puppet. If it did not, we're part of
an initial deployment, if it did exist then this is a redeploy and
in such case, if the config changed we will restart the service on
the bootstrap node.

We also cleanup the molecule steps as there are some redundancies there.

Testing protocol is in the parent change.

Related-Bug: #1942309

Change-Id: I08cff959d7d104ad0e8d833ed6ce06df8ff38b16