tripleo_ovs_dpdk: tripleo_ovs_dpdk_socket_memory is now optional
With ovs2.16, "dpdk-socket-limit" is not automatically set to the same
value as "dpdk-socket-mem".
dpdk-socket-mem is also optional as ovs will use hugepages as needed.
Yet, we dpdk-socket-mem is defined, the allocator may not use the
pre-allocated hugepages and use additional hugepages, that's why we
need to set dpdk-socket-limit as well.

Setting "tripleo_ovs_dpdk_socket_memory" keeps the behaviour
from ovs2.15.

RHBZ: 1985506
Change-Id: I5dad962a9e014380d90377a241ef090a61c873bd
Signed-off-by: Christophe Fontaine <cfontain@redhat.com>