Enforce root user at the beginning of container build
As long as we use a "standard" base image such as ubi8, we're already
root in the container when we run the different steps (SHELL, RUN, COPY
and so on).

But if we decide to inherit from another image that is actually setting
another value for USER (for instance, in TCIB, by passing
`tcib_user=foo'), we end up with permissions issues, preventing to
extend the inherited container.

This patch corrects this issue, enforcing the root user before anything
happens.

Change-Id: Ieb0677cae23384ee92f6a103dab8cd530591dd8d