bnr: Fixed the ERROR 1141 when users has fqdn
When the environment is deployed with TLS-e, the galera database has the user associates to these
hosts:
 - %
 - VIP
 - FQDN

The grants for the hosts % and VIP are populated. The FQDN ones are not populated.
So when the script to backup the grant of the users is executed, it shows an
irrelevant error like:

ERROR 1141 (42000) at line 1: There is no such grant defined for user "cinder"
on host "control.internal.api.example".

This patch backs up the grants of the users associate to:
 - %
 - localhost
 - IPV4
 - IPV6

Change-Id: I2b0c9a443cc1167a0183daf59ec2db3d4fc07d74
(cherry picked from commit 669b4f019296d478581ffde5bbc89dd8a0f24747)
(cherry picked from commit dc832f3416391902af5529b5b420936ef94fc044)