fix galaxy meta requirements
Also takes the opportunity to ignore "meta" directory from molecule
jobs. Since we apparently can't use "irrevelant_files" with "files",
we're using a negative look-ahead regexpi[1], to take all BUT "meta.*".

[1] https://www.regular-expressions.info/lookaround.html

NOTE: this patch also toggle tripleo-ansible-centos-8-role-addition to
non-voting in order to get this change merged. This job times out on a
regular basis, and prevents this needed patch to merge. Follow-up are
needed, for instance calling "pip-compile" as it's now done here:
https://review.opendev.org/c/openstack/tripleo-quickstart/+/807491

Closes-Bug: 1943965
Change-Id: I84d077c6daefdbeda4fda2d1f8fdd400c7437500
Signed-off-by: Kevin Carter <kecarter@redhat.com>