Don't use upper-constraints with molecule
We were always using the upper-constraints when creating tox
environments, however there is currently an issue with molecule and
openstack upper-constraints that is preventing it from being used.
Until molecule can be fixed to work with openstack upper-constraints,
lets drop this configuration. We are already adding the constraints to
the doc and release notes jobs.

https://github.com/ansible-community/molecule/pull/3244

Change-Id: I91ebaa059dc2451bba739a0a2de917e2f3c2aa21
Related-Bug: #1942704