Don't use upper-constraints with molecule
We were always using the upper-constraints when creating tox
environments, however there is currently an issue with molecule and
openstack upper-constraints that is preventing it from being used.
Until molecule can be fixed to work with openstack upper-constraints,
lets drop this configuration. We are already adding the constraints to
the doc and release notes jobs.

https://github.com/ansible-community/molecule/pull/3244

STABLE-Only: This also pins molecule to <3.5 as that requires updates
to the metadata structure.

Change-Id: I91ebaa059dc2451bba739a0a2de917e2f3c2aa21
Related-Bug: #1942704
(cherry picked from commit 949414244467e8436c07c2563643267adfb11a75)