Automate the /var/lib/ceph backup during recovery phase
To keep strong consistency for Ceph when Ceph nodes are collocated with
controllers, it is necessary to take a backup of the /var/lib/ceph
directory just before doing a recovery, restore the ReaR backup and then
extract the /var/lib/ceph backup over the (old) ceph data that is
outdated. With this simple action, the information lost is almost none.

This change creates two new scripts that are executed by ReaR during
restore that do exactly what is needed, eliminating the need for doing
it by hand.

Change-Id: Ib7c699b737980c5495cce2cf42f5847fda0f72f5