bnr Use replace instead of lineinfile and grouped similar tasks in a loop
Currently when the backup is executed and the history is kept,
there was two tasks that modify the configuration adding "timestamp"
to the name of the files of a couple of variables.

The modification of the configuration was done by lineinfile, but it
added the line whether exits or not. That's why the use of the replace
module instead of lineinfile.

This patch grouped these two tasks in one with a loop and also added
the variable OUTPUT_PREFIX_PXE to be configured in case of integration
with ironic.

Change-Id: I0260c8b5d1a6ccbddf1029ee9a3af35cb77fd525