Make Galera desync operation more resilient
Sometimes, after a hard reboot of a controller node the Galera database
is up&running but not yet able to receive requests via the mysql.sock
socket, even if it is already opened. It takes few seconds to get it
working, but during that time, we should keep retrying the desync
operation until it connects to the database.

Closes-Bug: #1945639

Change-Id: I71fe4c14195ebcf11912753425ea15bc679b14f9