Add tripleo_dnf_stream Ansible module.
One of the functionalities the dnf Ansible module misses is the single
disabling and disabling of DNF modules [0]. Currently only installation and
removal are supported.

The tripleo_dnf_stream module is based on the dnf.py Ansible core module [1]
(so it will be simpler to contribute with this functionality to upstream
Ansible core) and covers the enabling and disabling of one or multiple
MODULE:STREAM.

This functionality is required when performing a minor update, as we
need to make sure that a set of streams are enabled to provide with
the right packages [2]. In this situation, the installation of the
whole module is not really required as a full upgrade will run after.

The module uses the options name (MODULE:STREAM) and state:[disabled or
enabled]. Example:
    - name: Enable the container-tools:3.0
      tripleo_dnf_stream:
        - name: container-tools:3.0
          state: enabled

[0]: https://github.com/ansible/ansible/issues/64852
[1]: https://github.com/ansible/ansible/blob/devel/lib/ansible/modules/dnf.py
[2]: https://github.com/openstack/tripleo-heat-templates/blob/master/deployment/tripleo-packages/tripleo-packages-baremetal-puppet.yaml#L378-L381

Change-Id: I992935e1c478591f3c8af5044df8a625af9eba85
(cherry picked from commit 5212d407d9730dce68782d37f2045a672dce9146)
(cherry picked from commit c2d083b03c8314eeb1fdb606561e626fe89bffee)