Use service_net_map for ganesha_vip
Instead of if conditionals on hard-coded network names, use the
ServiceNetMap. GaneshaNetwork defaults to 'storage_nfs' with a
fallback to 'external' and then to 'ctlplane' in the ServiceNetMap.

This will allow custom network names to be used,
either via overriding GaneshaNetwork in ServiceNetMap or using the
service_net_map_replace in network data.

Related-Bug: #1946239
Depends-On: I5804621a3081209bbae63193c1e09601826f4275
Change-Id: Ic7743e7cd640703d166aecd97f59cbaf11c65500