Add the MTU size to ping to catch network issues
Many times users hit this issue where nic templates set jumbo
frames properly but switches are not properly configured and this
situation creates random issues hard to troubleshoot ranging from
a CEPH cluster not forming to swift not syncing properly.  Adding
an early test for MTU sizes in the ansible playbooks should fail
early and put us on a path to early solve this issue instead of
spending a week or more trying to figure out why CEPH batch LVM
is not working.

Change-Id: I9e56c512a563c4b04f398581e87846901fe021d6