bnr Add folder marker to avoid overwrote
Add a specific marker (directory name) to the blockinfile to avoid
deletion for any updates of the execution.

Currently, to configure the ironic integration backup. There is a
need of a backup data server (nfs) and the ironic PXE boot files
on the undercloud. If the backup data server is going to be on the
undercloud. We basically execute twice the nfs server configuration
with the data backup directory and the ironic configuration. The
way that we configure the nfs server, the blockinfile is overwritten.
So only the last configuration remains.

This patch aims to modify the blockinfile only when the directory
matches. For two different directories, it will create two
blockinfile blocks.

Change-Id: Ia30f4f27ee1a9cdd604d3c63d2d135fc471adc76
(cherry picked from commit 2f231ee71bf3d79257cf157917a433d7205fe93f)