Use shell module to generate keys
openssh_keypair ansible module was moved to community.crypto
collection from Ansible-Core 2.11, we can't use it now.
For using it in the product we either need to package and add to
deps and maintain the whole community.crypto module, either to
use 'shell' command to generate the keys.
Added "creates" so keys won't be overridden if they exists,
keeping the idempotency. [1]

[1] https://github.com/ansible-collections/community.crypto/blob/
91d98c44137cde53726bfa8f61898f3c28041e20/plugins/module_utils/openssh/backends/common.py#L188

Change-Id: Ib9a31518ee9408d89abff4c4eb18a7cfe243fb58