bnr - Remove the filter to backup the root grants [mysql]
Currently the root user is rule out of the mysql grants backup when
the backup of the mysql database is done.
This means when the mysql backup is restored the operator needs
to set back the mysql root password again if they want to use the
same process.

Closes-Bug: #1952099
Change-Id: Ia1ecbe53605f06b233df6406a0b9fe23e8ba22a8
(cherry picked from commit b50212faa8ecb297d6a569f8fbbb3cf9769d77f1)