Support a custom ceph_crush_hierarchy in the Ceph spec
This patch brings a new feature: it's now possible to build
a customized crush hierarchy associated to the OSD nodes.
The ceph_spec library now supports the 'location' parameter,
which is a dictionary containing the OSD position within the
crush hierarchy.
The 'ceph_bootstrap' ansible module is updated to accept the
new ceph_crush_hierarchy parameter, which is supposed to be
passed as extra_var to the cephadm playbook.

Change-Id: I46c7904d15ef48a45966eb2dac065593a3dc21cb