Turn crush_hierarchy path into var content
Like we've done for the osd_spec, the tripleclient command provides
a path file where the crush hierarchy is defined.
This task looks for the content of the hierarchy file, turning it
into a var content that is passed to the ceph_spec_bootstrap module.

Change-Id: I09242a8cf6286dc5b9263afb2e22b8d5cdad49f1
(cherry picked from commit 551baec5741d67e8bcf953b5b33a692b8f3da513)