Disable image-serve MultiViews
The MultiViews configuration for manifests is causing the wrong
manifest data being returned for some tags.  This change removes the
MultiViews configuration and moves the configuration we used to put into
an .htaccess file into the image serve apache configuration.  The
manifest requests should no longer trigger the .htaccess configuration
but the .htaccess for the actual metadata file should still be returned
with this change.

Change-Id: Ifd86b50e308bca0b8fd4ddb870b289dde8dbe046
Closes-Bug: #1953198