Remove redundant default route from ci nic-config
The current nic-config is laying redundant default routes
for some roles (Controller/compute-dvr), logs [1].

* On ctlplane
* On external network - A route for 0.0.0.0/0 is a default route

With the default route on external network overriding the
ctlplane one[2]. Having redundant default route is confusing add
there should be only one default route.

With this patch, we are move the default route on external n/w
into the condition only when external network is defined.
Also add a inverted condition around the routes on the ctlplane
interface, to ensure we get either not both.

More discussion around this on [2]

[0] https://logserver.rdoproject.org/29/818529/3/openstack-check/tripleo-ci-centos-8-ovb-3ctlr_1comp-featureset001/ff2a816/logs/overcloud-controller-0/etc/os-net-config/config.yaml.txt.gz
[1] https://logserver.rdoproject.org/29/818529/3/openstack-check/tripleo-ci-centos-8-ovb-3ctlr_1comp-featureset001/ff2a816/logs/overcloud-controller-0/var/log/extra/network.txt.gz
[2] https://review.opendev.org/c/openstack/tripleo-ansible/+/818529/3#message-2a8981e1803ea304e61a5a21ec85fb5ae4cb6073

Change-Id: I7b5f36012240aa8dba62857a873e69543fc61a8b