Create nova flavor for Octavia
The nova flavor for Octavia was previously created from puppet-octavia
but it didn't offer a convenient way to select the type of the API
endpoint (it uses public API endpoint by default while some deployments
don't have access to the external network from their controllers).

Moving the creation of the flavor to a tripleo-ansible role helps us to
use the existing configuration (the roles use the internal endpoints)
and centralizes the creation of all the resources required by Octavia in
a single place.

Change-Id: I6cbb9a50f4bf952e927f6eba7fdb993cfbb9cb72