Add tripleo_cephadm_verbose condition to post.yaml
The final step of the cephadm playbook calls post.yaml, which is able
to print the status of the cluster and export it as a ceph compatible
spec retrieved from the Ceph cluster. However, in a tls-everywhere
context, we should be able to stop logging the spec as it contains
sensitive info (the priv key).
This change just adds the tripleo_cephadm_verbose condition to control
what is printed.

Change-Id: I7bef5922d75114e4bd1924b52a382e61d7027d2d