Ensure httpd is allowed to listen on the port
Apparently, starting cs9, httpd isn't allowed to listen on the default
image_serve port (8787) anymore.

While tripleo-heat-templates modifies the selinux type for this port,
there are some cases where we run this tripleo_image_serve role from
outside the deploy, and we'll face denials for that port.

Since we already manage some of the SELinux configs for the directory,
we can also manage the port.

We also need to add a new dependency in order to run molecule.

Change-Id: I4953aeeef1588538448696ffbb463f9da3ff0571