Add mds, nfs and ingress daemons to deployed ceph
This change moves the logic of deploying mds/ganesha and the related
ingress resource to deployed ceph.  By passing a new option to deployed
ceph is possible to build the mds, nfs and ingress spec files and apply
them against the ceph cluster, as well as configuring the manila pools
and cephfs.

Change-Id: Ic0b26d7ee505ca84129782f58f9d14f815467854