Provision VIPs for ceph related services
This change adds a new task within cli-deployed-ceph that will be
triggered by tripleoclient when --ceph-vip option is set.
The extra_var input file defines the mapping between the ceph services
and their network, and produces (as output) the VIPs associated to the
specified daemons.
It uses the improved version of "tripleo_service_vip" ansible module
that now supports an output path instead only a playbook dir, and is
able to return a payload (the generated data map) that can be used by
other tasks: this ensures we have more control over the provided input
and output.

Change-Id: I5c2e5caa1579af5cf206633447b92e8fdffdda16