Rename variables to keep being the same as in wallaby
Some variables needed to be changed to adapt to the new process of doing
restores using PXE/Ironic. Even if this feature is not going back to v,
u and t, we would want to keep the variables that do the same called the
same among all versions of the product. This change renames some
variables to achieve this.

Change-Id: I79d8011d128357282fd1a847a24fb11971eb1487