tripleo-bootstrap - check packeges fact befor install
When repository files are present, but the ctlplane is not on a
routed network. (I.e the node cannot connect to repositories.)
The ansible built in `package` will fail despite the fact that
the packages is already installed.

This change adds a task to check that the packages required are
isntalled using `rpm --query --whatprovides <PGK_LIST>`.
`--whatprovides` is used to find out if a wrapper rpm for OVS with
custom versioning scheme is installed. If a wrapper is installed
'openvswitch' is appended to the list of installed packages.

Related: RHBZ#2048134
Closes-Bug: #1959731
Change-Id: Ie448e8f795325cc3ae357fae32f3a47406101f25
(cherry picked from commit 37fac649388e0bf39326bdaaefa5a104b6ef264c)
(cherry picked from commit 7b002dec7dd32322023487f78cf7944aee7b1c46)