tuned - check packages before install
When repository files are present, but the ctlplane is not on a
routed network. (I.e the node cannot connect to repositories.)
The ansible built in `package` will fail despite the fact that
the packages is already installed.

This change adds a task to check that the packages required are
installed using `rpm --query --whatprovides <PGK_LIST>`. Pacage
installation task is skipped if package already installed.

Related: RHBZ#2048134
Related-Bug: #1959731
Change-Id: Ifcbde65af99039a261bc34a813bddfb29f63c1e1