[only train] BnR set default value for variable
The variable tripleo_backup_and_restore_hide_sensitive_logs is used
on the role but it isn't initialized. When the variable is not defined
and the galera desyncronization fails, the task fails and do not
repeat the amount of attemps that it should've tried.

This patch only defines the variable to true, so the logs are
hidden by default.

Change-Id: I34cdc2045cf668f70f56a5e2e807754125c9be99