Modify cert gen to make it work under fips
Some openssl command line utilities (e.g. openssl req) accept a
supplementary config file as an option. This config file should
contain only supplementary sections. If you use the main
configuration file as supplementary, it may cause strange errors
because openssl will try to activate the providers twice. This
behaviour existed for a long time but before the 3.0 version
it caused problems less often, and it appears to fail consistently
under FIPS.

In this commit, we create a config file with just those configs
that are necessary.

Change-Id: Iecf3f3e20fa285563b23a7aff7deaa00616aca51