Add pre-provisioned nodes to unprovision confirmation
In order for tripleoclient to properly display a confirmation about
unprovisioning network ports for pre-provisioned nodes, the
pre-provisioned node data must be added to the unprovision confirmation.

The output structure written to the unprovision confirmation file is
changed from a list to a dict so that the provisioned vs pre-provisioned
nodes can be distinguished. This change would break tripleoclient,
however the associated tripeloclient depends-on patch adds backwards
compatibility to check for a list or dict.

Signed-off-by: James Slagle <jslagle@redhat.com>
Depends-On: Id7ebc59948cc6ad6c2270304c35ab0475259a6c0
Change-Id: Id377b2c7ed973a7e2365b33dba2c1d59a59e26b9