Support both inventory and metal file in ceph_spec_bootstrap
The deployed metal file comes from the output of 'openstack
overcloud node provision'. If this file's DeployedSeverPortMap
is empty, we cannot map hostnames to their provisioning IPs.
This information is in the inventory however so, if we have
an inventory, then use it to complete the mapping. We do this
by adding an option to ceph_spec_bootstrap to use both the
inventory and the metal file as a source to build the spec.

The cli-deployed-ceph.yaml playbook was designed to work with
an inventory containing only the undercloud and a populated
DeployedSeverPortMap. The ceph_spec_bootstrap module still
supports this behavior though the playbook now passes the
"both" option and thus requires an inventory with all hosts.
Such an inventory is genereated by 'openstack overcloud node
provision'. When the "both" case is used, it is assumed the
inventory does not have the ceph_mon group so the playbook
still dynamically expands its inventory to add it.

Closes-Bug: 1962226
Change-Id: I17f4303636b08d2be2d0f6e248242ad53c5fddf1